Update README.md
This project is a simple address book meant to store users data for future use. The user simply enters their information into the system, and then the data is stored locally on the hosts computer. The file holds all the users data so that it can be read at a later date.